Added mode support
Three new modes for item selection is introduced namely Simple,
Single Selection and Multi Selection.

In Single Selection, the items under a parent can be selected in a
way similar to radio buttons. In Multi Selection, items can be
selected in a checkbox manner. Simple Selection on the other hand
is the default method available previously.