Replaced printf usage with own print function
A print function which mimics the printf function is written to
replace the usage of printf. The function is implemented in
apptree_io.c. The function would require that a function which
redirects all output to harware first be binded to apptree_io.
The reason for replacing the printf is to make the initialization
of the apptree more simpler and more intuitive. For example, the
user does not have to set up and redirect the standard io. Whereas,
he now only has to redirect the output of the apptree. This also
makes the functionality of the apptree more transparent.

The read input function in apptree.c is also moved to apptree_io
for uniformity. This means that all io is now effectively handled
by apptree_io.